Liquidity query to correctly use basin
