+++ src/queries/bean/AdvancedChart.graphql
@@ -1,5 +1,5 @@
 query BeanAdvancedChart($from: Int, $to: Int) {
-  seasons(first: 1000, orderBy: season, orderDirection: desc, where: {  season_gte: $from, season_lte: $to } ) {
+  seasons(first: 1000, orderBy: season, orderDirection: desc, where: { season_gte: $from, season_lte: $to } ) {
     id
     timestamp
     beanHourlySnapshot {
