--- src/queries/bean/seasonal/SeasonalSeasonPools.graphql
@@ -1,18 +1,0 @@
-query BeanSeasonalSeasonPools($from: Int, $to: Int) {
-  seasons(
-    where: { season_gte: $from, season_lte: $to }
-    first: 1000
-    orderBy: season
-    orderDirection: asc
-  ) {
-    id
-    season
-    poolHourlySnapshots {
-      pool {
-        id
-      }
-      liquidityUSD
-      createdTimestamp
-    }
-  }
-}
